+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.8] - 2026-07-15
+
+### Fixed
+- Energy flow sensors now have correct units (% for SOC, W/kW for power).
+- Removed queryPlants API calls that don't exist in ValueClouds API.
+
 ## [2.3.7] - 2026-07-15
 
 ### Fixed